fix(functions): restore sandbox account context lookup
