+++ packages/sandbox/lib/index.ts
@@ -1,4 +1,3 @@
-export * from './api-keys/account-context.js';
 export * from './api-keys/sandbox-api-key.service.js';
 export { default as sandboxApiKeyService } from './api-keys/sandbox-api-key.service.js';
 export * from './functions/dryruns.service.js';
